fix program2list

+++ src/Simulation.hs
@@ -1,10 +1,10 @@
 {-# LANGUAGE RecordWildCards #-}
 module Simulation where
-    
+
 import Instructions
 
-data State = State { register :: [Int], 
-                     instructionCounter :: Int } 
+data State = State { register :: [Int],
+                     instructionCounter :: Int }
 
 -- Show który wypisuje pierwszy rejestr gdzie znajduje się Wynik
 instance Show State where
@@ -13,20 +13,20 @@ instance Show State where
 
 --Alternatywny show którego używałem do testowania wypisuje n rejestrów. Przydatne przy obliczeniach krok po kroku
 showState :: Int -> State -> String
-showState n state@State{..} = "Rejestr: " ++ show (take n register) ++ " Licznik: " ++ show instructionCounter 
+showState n state@State{..} = "Rejestr: " ++ show (take n register) ++ " Licznik: " ++ show instructionCounter
 
 -- Funkcja służy do generwoania stanu pustego
-emptyState :: State 
+emptyState :: State
 emptyState = State {register = repeat 0, instructionCounter = 0}
 
 -- Funkcja służy do generwania stanu początkowego dla danych wejściowych w formie listy
-initialState :: [Int] -> State 
+initialState :: [Int] -> State
 initialState x = State {register = x ++ repeat 0, instructionCounter = 0}
 
 -- Funkcja INC 
 inc :: Int -> State -> State
 inc i state@State{..} = state{register = updateList i (+1) register, instructionCounter = instructionCounter + 1}
-  where 
+  where
     updateList n f xs = take n xs ++ [f (xs !! n)] ++ drop (n + 1) xs
 
 -- Funkcja DEC
@@ -60,5 +60,5 @@ runProgramStepByStep list state@State{..}
 
 program2list :: Program -> [State -> State]
 program2list [] = []
-program2list ((Inc x):xs) = (inc x):program2list xs
-program2list ((Dec x y):xs) = (dec x y):program2list xs
+program2list ((Inc x):xs) = inc (fromIntegral x):program2list xs
+program2list ((Dec x y):xs) = dec (fromIntegral x) (fromIntegral y):program2list xs
